fix(pool): use time.Duration for idleTimeout to prevent int overflow on 32-bit systems

+++ pkg/testutil/testutil.go
@@ -25,7 +25,7 @@ func NewTestPoolManager(t *testing.T) *pool.Manager {
 		},
 		1,
 		10,
-		int(1*time.Minute),
+		1*time.Minute,
 		false,
 	)
 	require.NoError(t, err)
